Remove deprecated array column type from entities

+++ src/Command/MigrateCommand.php
@@ -389,6 +389,11 @@ class MigrateCommand extends Command
                 return 2;
             }
 
+            if (!$this->helper->capsule( 'doctrine:migrations:version "DoctrineMigrations\Version20250830134338" --add', $output )) {
+                $output->writeln("<error>Unable to create schema.</error>");
+                return 2;
+            }
+
             if (!$this->helper->capsule( 'doctrine:fixtures:load --append', $output )) {
                 $output->writeln("<error>Unable to update fixtures.</error>");
                 return 3;
@@ -548,7 +553,7 @@ class MigrateCommand extends Command
 
                     $source = "{$this->param->get('kernel.project_dir')}/migrations";
                     foreach (scandir( $source ) as $file)
-                        if ($file && $file[0] !== '.') {
+                        if ($file && $file[0] !== '.' && $file !== 'Version20250830134338.php') {
                             $output->write("\tDeleting \"<comment>{$file}</comment>\"... ");
                             unlink( "$source/$file" );
                             $output->writeln('<info>Ok!</info>');
@@ -971,7 +976,7 @@ class MigrateCommand extends Command
                     if ($spawn_zone) {
                         $output->writeln("Spawning <info>{$spawning_ruin->getLabel()}</info> at <info>{$spawn_zone->getX()} / {$spawn_zone->getY()}</info>");
                         $spawn_zone->setPrototype($spawning_ruin);
-                        
+
                         $this->maze->setTargetZone($spawn_zone);
                         $spawn_zone->setExplorableFloors($this->conf->getTownConfiguration($town)->get(TownSetting::ERuinSpaceFloors));
                         $this->maze->createField();
